Added pins and zoom into pins
Summary: Pins appear on the map in the locations corresponding to the actual placs on the recommendations list. Also, once the pins are added, the map zooms in (such that all pins are shown).

Test Plan: N/A

Reviewers: madjangba, kiira, kychow

Reviewed By: madjangba

Differential Revision: https://phabricator.intern.facebook.com/D8984260

Signature: 8984260:1532547649:126f39890433b4d03fc167019cd7d9dcd60022d6

+++ app/src/main/java/com/kychow/jayjoska/models/Place.java
@@ -14,8 +14,8 @@ public class Place {
     private String businessURL;
     private long rating;
     private long distance;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     
     /* Takes in a JSONObject and parses for the corresponding data*/
     public static Place fromJSON(JSONObject jsonObject) throws JSONException {
@@ -33,8 +33,8 @@ public class Place {
         place.distance = jsonObject.getLong("distance");
         //access Coordinates and grab latitude and longitude
         JSONObject coord = jsonObject.getJSONObject("coordinates");
-        place.latitude = coord.getLong("latitude");
-        place.longitude = coord.getLong("longitude");
+        place.latitude = coord.getDouble("latitude");
+        place.longitude = coord.getDouble("longitude");
 
         return place;
     }
@@ -63,11 +63,11 @@ public class Place {
         return distance;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 }
